[fix] Fix godfrey's comments
