+++ flink-table/flink-table-planner-blink/src/main/java/org/apache/flink/table/planner/sinks/BatchSelectTableSink.java
@@ -18,76 +18,25 @@
 
 package org.apache.flink.table.planner.sinks;
 
-import org.apache.flink.api.common.ExecutionConfig;
-import org.apache.flink.api.common.typeutils.TypeSerializer;
-import org.apache.flink.core.execution.JobClient;
 import org.apache.flink.streaming.api.datastream.DataStream;
 import org.apache.flink.streaming.api.datastream.DataStreamSink;
-import org.apache.flink.streaming.api.operators.collect.CollectResultIterator;
-import org.apache.flink.streaming.api.operators.collect.CollectSinkOperator;
-import org.apache.flink.streaming.api.operators.collect.CollectSinkOperatorFactory;
-import org.apache.flink.streaming.api.operators.collect.CollectStreamSink;
 import org.apache.flink.table.api.TableSchema;
 import org.apache.flink.table.api.internal.SelectTableSink;
-import org.apache.flink.table.runtime.types.TypeInfoDataTypeConverter;
 import org.apache.flink.table.sinks.StreamTableSink;
-import org.apache.flink.table.types.DataType;
 import org.apache.flink.types.Row;
 
-import java.util.Iterator;
-import java.util.UUID;
 
 /**
  * A {@link SelectTableSink} for batch select job.
- *
- * <p><strong>NOTES:</strong> This is a temporary solution,
- * once FLINK-14807 is finished, the implementation should be changed.
  */
-public class BatchSelectTableSink implements StreamTableSink<Row>, SelectTableSink {
+public class BatchSelectTableSink extends SelectTableSinkBase implements StreamTableSink<Row> {
 
-	private final TableSchema tableSchema;
-	private final CollectSinkOperatorFactory<Row> factory;
-	private final CollectResultIterator<Row> iterator;
-
-	@SuppressWarnings("unchecked")
 	public BatchSelectTableSink(TableSchema tableSchema) {
-		this.tableSchema = SelectTableSinkSchemaConverter.convertTimeAttributeToRegularTimestamp(
-				SelectTableSinkSchemaConverter.changeDefaultConversionClass(tableSchema));
-
-		TypeSerializer<Row> typeSerializer = (TypeSerializer<Row>) TypeInfoDataTypeConverter
-				.fromDataTypeToTypeInfo(this.tableSchema.toRowDataType())
-				.createSerializer(new ExecutionConfig());
-		String accumulatorName = "tableResultCollect_" + UUID.randomUUID();
-
-		this.factory = new CollectSinkOperatorFactory<>(typeSerializer, accumulatorName);
-		CollectSinkOperator<Row> operator = (CollectSinkOperator<Row>) factory.getOperator();
-		this.iterator = new CollectResultIterator<>(operator.getOperatorIdFuture(), typeSerializer, accumulatorName);
-	}
-
-	@Override
-	public DataType getConsumedDataType() {
-		return tableSchema.toRowDataType();
-	}
-
-	@Override
-	public TableSchema getTableSchema() {
-		return tableSchema;
+		super(tableSchema);
 	}
 
 	@Override
 	public DataStreamSink<?> consumeDataStream(DataStream<Row> dataStream) {
-		CollectStreamSink<Row> sink = new CollectStreamSink<>(dataStream, factory);
-		dataStream.getExecutionEnvironment().addOperator(sink.getTransformation());
-		return sink.name("Batch select table sink");
-	}
-
-	@Override
-	public void setJobClient(JobClient jobClient) {
-		iterator.setJobClient(jobClient);
-	}
-
-	@Override
-	public Iterator<Row> getResultIterator() {
-		return iterator;
+		return super.consumeDataStream(dataStream);
 	}
 }
