+++ flink-table/flink-table-planner-blink/src/main/java/org/apache/flink/table/planner/sinks/StreamSelectTableSink.java
@@ -18,79 +18,28 @@
 
 package org.apache.flink.table.planner.sinks;
 
-import org.apache.flink.api.common.ExecutionConfig;
-import org.apache.flink.api.common.typeutils.TypeSerializer;
-import org.apache.flink.core.execution.JobClient;
 import org.apache.flink.streaming.api.datastream.DataStream;
 import org.apache.flink.streaming.api.datastream.DataStreamSink;
-import org.apache.flink.streaming.api.operators.collect.CollectResultIterator;
-import org.apache.flink.streaming.api.operators.collect.CollectSinkOperator;
-import org.apache.flink.streaming.api.operators.collect.CollectSinkOperatorFactory;
-import org.apache.flink.streaming.api.operators.collect.CollectStreamSink;
 import org.apache.flink.table.api.TableSchema;
 import org.apache.flink.table.api.internal.SelectTableSink;
-import org.apache.flink.table.runtime.types.TypeInfoDataTypeConverter;
 import org.apache.flink.table.sinks.AppendStreamTableSink;
-import org.apache.flink.table.types.DataType;
 import org.apache.flink.types.Row;
-
-import java.util.Iterator;
-import java.util.UUID;
 
 /**
  * A {@link SelectTableSink} for streaming select job.
  *
- * <p><strong>NOTES:</strong> This is a temporary solution,
- * once FLINK-14807 is finished, the implementation should be changed.
+ * <p><strong>NOTE:</strong>
  * Currently, only insert changes (AppendStreamTableSink) is supported.
  * Once FLINK-16998 is finished, all kinds of changes will be supported.
  */
-public class StreamSelectTableSink implements AppendStreamTableSink<Row>, SelectTableSink {
+public class StreamSelectTableSink extends SelectTableSinkBase implements AppendStreamTableSink<Row> {
 
-	private final TableSchema tableSchema;
-	private final CollectSinkOperatorFactory<Row> factory;
-	private final CollectResultIterator<Row> iterator;
-
-	@SuppressWarnings("unchecked")
 	public StreamSelectTableSink(TableSchema tableSchema) {
-		this.tableSchema = SelectTableSinkSchemaConverter.convertTimeAttributeToRegularTimestamp(
-				SelectTableSinkSchemaConverter.changeDefaultConversionClass(tableSchema));
-
-		TypeSerializer<Row> typeSerializer = (TypeSerializer<Row>) TypeInfoDataTypeConverter
-			.fromDataTypeToTypeInfo(this.tableSchema.toRowDataType())
-			.createSerializer(new ExecutionConfig());
-		String accumulatorName = "tableResultCollect_" + UUID.randomUUID();
-
-		this.factory = new CollectSinkOperatorFactory<>(typeSerializer, accumulatorName);
-		CollectSinkOperator<Row> operator = (CollectSinkOperator<Row>) factory.getOperator();
-		this.iterator = new CollectResultIterator<>(operator.getOperatorIdFuture(), typeSerializer, accumulatorName);
-	}
-
-	@Override
-	public DataType getConsumedDataType() {
-		return tableSchema.toRowDataType();
-	}
-
-	@Override
-	public TableSchema getTableSchema() {
-		return tableSchema;
+		super(tableSchema);
 	}
 
 	@Override
 	public DataStreamSink<?> consumeDataStream(DataStream<Row> dataStream) {
-		CollectStreamSink<Row> sink = new CollectStreamSink<>(dataStream, factory);
-		dataStream.getExecutionEnvironment().addOperator(sink.getTransformation());
-		return sink.name("Streaming select table sink");
-	}
-
-	@Override
-	public void setJobClient(JobClient jobClient) {
-		iterator.setJobClient(jobClient);
-	}
-
-	@Override
-	public Iterator<Row> getResultIterator() {
-		return iterator;
+		return super.consumeDataStream(dataStream);
 	}
 }
-
